requirements/flask/requirements.txt
Flask
Flask-MySQLdb
mysql-connector-python
python-dotenv
requests
yfinance